% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolutizePath.R
\name{absolutizePath}
\alias{absolutizePath}
\title{Make an absolute file path}
\usage{
absolutizePath(path)
}
\arguments{
\item{path}{String containing an absolute or relative file path.}
}
\value{
An absolute file path corresponding to \code{path}, or \code{path} itself if it was already absolute.
}
\description{
Create an absolute file path from a relative file path.
All processing is purely lexical; the path itself does not have to exist on the filesystem.
}
\examples{
absolutizePath("alpha")
absolutizePath("/alpha/bravo")

}
\author{
Aaron Lun
}
