% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRedirection.R
\name{.createRedirection}
\alias{.createRedirection}
\title{Create a redirection file}
\usage{
.createRedirection(dir, src, dest)
}
\arguments{
\item{dir}{String containing the path to the staging directory.}

\item{src}{String containing the source path relative to \code{dir}.}

\item{dest}{String containing the destination path relative to \code{dir}.}
}
\value{
A list of metadata that can be processed by \code{\link{.writeMetadata}}.
}
\description{
Create a redirection to another path in the same staging directory.
This is useful for creating short-hand aliases for resources that have inconveniently long paths.
}
\details{
\code{src} should not correspond to an existing file inside \code{dir}.
This avoids ambiguity when attempting to load \code{src} via \code{\link{acquireFile}}.
Otherwise, it would be unclear as to whether the user wants the file at \code{src} or the redirection target \code{dest}.
}
\author{
Aaron Lun
}
