% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDirectory.R
\name{listDirectory}
\alias{listDirectory}
\alias{listLocalObjects}
\title{List staged objects in a directory}
\usage{
listDirectory(dir, ignore.children = TRUE)
}
\arguments{
\item{dir}{String containing a path to a staging directory.}

\item{ignore.children}{Logical scalar indicating whether to ignore metadata for child objects.}
}
\value{
Named list where each entry is itself a list containing the metadata for an object.
The name of the entry is the path inside \code{dir} to the object.

If \code{ignore.children=TRUE}, metadata is only returned for non-child objects or redirections.
}
\description{
List all objects in a staging directory by loading their metadata.
}
\examples{
tmp <- tempfile()
dir.create(tmp)

library(S4Vectors)
df <- DataFrame(A=1:10, B=LETTERS[1:10])
meta <- stageObject(df, tmp, path="whee")
writeMetadata(meta, tmp)

ll <- list(A=1, B=LETTERS, C=DataFrame(X=1:5))
meta <- stageObject(ll, tmp, path="stuff")
writeMetadata(meta, tmp)

redirect <- createRedirection(tmp, "whoop", "whee/simple.csv.gz")
writeMetadata(redirect, tmp)

all.meta <- listDirectory(tmp)
names(all.meta) 

}
\author{
Aaron Lun
}
