% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAtomicVector.R
\name{loadAtomicVector}
\alias{loadAtomicVector}
\title{Load an atomic vector}
\usage{
loadAtomicVector(info, project, ...)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}

\item{...}{Further arguments, ignored.}
}
\value{
The vector described by \code{info}, possibly with names.
}
\description{
Load a simple vector consisting of atomic elements from file.
}
\examples{
tmp <- tempfile()
dir.create(tmp)
meta <- stageObject(setNames(runif(26), letters), tmp, path="bar")
loadAtomicVector(meta, tmp)

}
\seealso{
\code{"\link{stageObject,integer-method}"}, for one of the staging methods.
}
\author{
Aaron Lun
}
