% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBaseFactor.R
\name{loadBaseFactor}
\alias{loadBaseFactor}
\title{Load a factor}
\usage{
loadBaseFactor(info, project, ...)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}

\item{...}{Further arguments, ignored.}
}
\value{
The vector described by \code{info}, possibly with names.
}
\description{
Load a base R \link{factor} from file.
}
\examples{
tmp <- tempfile()
dir.create(tmp)
meta <- stageObject(factor(letters[1:10], letters), tmp, path="bar")
loadBaseFactor(meta, tmp)

}
\seealso{
\code{"\link{stageObject,factor-method}"}, for the staging method.
}
\author{
Aaron Lun
}
