% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBaseList.R
\name{loadBaseList}
\alias{loadBaseList}
\title{Load a base list}
\usage{
loadBaseList(info, project, parallel = TRUE)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}

\item{parallel}{Whether to perform reading and parsing in parallel for greater speed.
Only relevant for lists stored in the JSON format.}
}
\value{
The list described by \code{info}.
}
\description{
Load a \link{list} from file, possibly containing complex entries.
}
\details{
This function effectively reverses the behavior of \code{"\link{stageObject,list-method}"}, loading the \link{list} back into memory from the JSON file.
Atomic vectors, arrays and data frames are loaded directly while complex values are loaded by calling the appropriate loading function.
}
\examples{
library(S4Vectors)
ll <- list(A=1, B=LETTERS, C=DataFrame(X=letters))

# First staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(ll, tmp, path="stuff")

# And now loading it:
loadBaseList(info, tmp)

}
\seealso{
\code{"\link{stageObject,list-method}"}, for the staging method.
}
\author{
Aaron Lun
}
