% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFrameFactor.R
\name{loadDataFrameFactor}
\alias{loadDataFrameFactor}
\title{Load an DataFrame factor}
\usage{
loadDataFrameFactor(info, project)
}
\arguments{
\item{info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{DataFrameFactor} described by \code{info}.
}
\description{
Load a \linkS4class{Factor} where the levels are rows of a \linkS4class{DataFrame}.
}
\examples{
library(S4Vectors)
df <- DataFrame(X=LETTERS[1:5], Y=1:5)
out <- DataFrameFactor(df[sample(5, 100, replace=TRUE),,drop=FALSE])

# Staging the object:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(out, tmp, path="test")

# Loading it back in:
loadDataFrameFactor(info, tmp)

}
\seealso{
\code{"\link{stageObject,DataFrameFactor-method}"}, for the staging method.
}
\author{
Aaron Lun
}
