% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBaseList.R
\name{readBaseList}
\alias{readBaseList}
\alias{loadBaseList}
\alias{validateBaseList}
\title{Read a base list from disk}
\usage{
readBaseList(path, list.parallel = TRUE, ...)

validateBaseList(path)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with the list method for \code{\link{stageObject}}.}

\item{list.parallel}{Whether to perform reading and parsing in parallel for greater speed.
Only relevant for lists stored in the JSON format.}

\item{...}{Further arguments to be passed to \code{\link{altReadObject}} for complex child objects.}
}
\value{
For \code{readBaseList}, the list represented by \code{path}.

For \code{validateBaseList}, an error is raised if the on-disk representation is invalid.
Otherwise, \code{NULL} is silently returned.
}
\description{
Read a \link{list} from its on-disk representation.
}
\examples{
library(S4Vectors)
ll <- list(A=1, B=LETTERS, C=DataFrame(X=letters))

tmp <- tempfile()
saveObject(ll, tmp)
readBaseList(tmp)

}
\seealso{
\code{"\link{stageObject,list-method}"}, for the staging method.
}
\author{
Aaron Lun
}
