% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLocalObject.R
\name{readLocalObject}
\alias{readLocalObject}
\alias{saveLocalObject}
\alias{listLocalObjects}
\title{Convenience helpers for handling local directories}
\usage{
readLocalObject(dir, path, ...)

saveLocalObject(x, dir, path, ...)

listLocalObjects(dir, includeChildren = FALSE)
}
\arguments{
\item{dir}{String containing a path to the directory.}

\item{path}{String containing a relative path to the object of interest inside \code{dir}.}

\item{x}{Object to be saved.}

\item{includeChildren}{Logical scalar indicating whether child objects should be returned.}
}
\value{
For \code{readLocalObject}, the object at \code{path}.

For \code{saveLocalObject}, the object is saved to \code{path} inside \code{dir}.
All necessary directories are created if they are not already present.
A \code{NULL} is returned invisibly.

For \code{listLocalObjects}, a named list where each entry corresponds to an object and is named after the path to that object inside \code{dir}.
The value of each element is another list containing the metadata for its corresponding object.
}
\description{
Read, write and list objects from a local staging directory.
These are just convenience wrappers around functions like \code{\link{loadObject}}, \code{\link{stageObject}} and \code{\link{.writeMetadata}}.
}
\examples{
local <- tempfile()

# Creating a slightly complicated object:
library(S4Vectors)
df <- DataFrame(A=1:10, B=LETTERS[1:10])
df$C <- DataFrame(D=letters[1:10], E=runif(10))

# Saving it:
saveLocalObject(df, local, "FOOBAR")

# Reading it back:
readLocalObject(local, "FOOBAR")

# Listing all available objects:
names(listLocalObjects(local))
names(listLocalObjects(local, includeChildren=TRUE))
}
\author{
Aaron Lun
}
