% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreMetadata.R
\name{restoreMetadata}
\alias{restoreMetadata}
\alias{.restoreMetadata}
\title{Restore R-level metadata}
\usage{
restoreMetadata(x, mcol.data, meta.data, project)
}
\arguments{
\item{x}{An \linkS4class{Vector} or \linkS4class{Annotated} object.}

\item{mcol.data}{List of metadata specifying the \code{resource} location of the \code{\link{mcols}} files.
This can also be \code{NULL}, in which case no \code{mcols} are restored.}

\item{meta.data}{List of metadata specifying the \code{resource} location of the \code{\link{metadata}} files.
This can also be \code{NULL}, in which case no \code{metadata} are restored.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
\code{x} is returned, possibly with \code{mcols} and \code{metadata} added to it.
}
\description{
Restore \code{\link{metadata}} and \code{\link{mcols}} for a \linkS4class{Annotated} or \linkS4class{Vector} object, respectively.
This is typically used inside loading functions for concrete subclasses.
}
\seealso{
\code{\link{.processMetadata}}, which does the staging.
}
\author{
Aaron Lun
}
