% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveBaseList.R
\name{saveObject,list-method}
\alias{saveObject,list-method}
\alias{stageObject,List-method}
\title{Save a base list to disk}
\usage{
\S4method{saveObject}{list}(x, path, format = saveBaseListFormat(), ...)

\S4method{stageObject}{List}(x, dir, path, child = FALSE, fname = "list")
}
\arguments{
\item{x}{A Bioconductor object of the specified class.}

\item{path}{String containing the path to a directory in which to save \code{x}.}

\item{format}{String specifying the format in which to save the list.}

\item{...}{Further arguments, passed to \code{\link{altSaveObject}} for complex child objects.}

\item{A}{list or \linkS4class{List}.}
}
\value{
\code{x} is saved inside \code{dir}.
}
\description{
Save a \link{list} or \linkS4class{List} to a JSON or HDF5 file, with extra files created for any of the more complex list elements (e.g., DataFrames, arrays). 
This uses the \href{https://github.com/LTLA/uzuki2}{uzuki2} specification to ensure that appropriate types are declared.
}
\section{File formats}{

If \code{format="json.gz"}, the list is saved to a Gzip-compressed JSON file (the default).
This is an easily parsed format with low storage overhead.

If \code{format="hdf5"}, \code{x} is saved into a HDF5 file instead.
This format is most useful for random access and for preserving the precision of numerical data.
}

\examples{
library(S4Vectors)
ll <- list(A=1, B=LETTERS, C=DataFrame(X=1:5))

tmp <- tempfile()
saveObject(ll, tmp)
list.files(tmp, recursive=TRUE)

}
\seealso{
\url{https://github.com/LTLA/uzuki2} for the specification.
}
\author{
Aaron Lun
}
