% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageBaseFactor.R
\name{stageObject,factor-method}
\alias{stageObject,factor-method}
\title{Stage factors}
\usage{
\S4method{stageObject}{factor}(x, dir, path, simplified = FALSE, child = FALSE, ...)
}
\arguments{
\item{x}{Any of the assorted simple vector types.}

\item{...}{Further arguments that are ignored.}
}
\value{
A named list containing the metadata for \code{x}.
\code{x} itself is written to a file inside \code{path}, along with the various levels.
}
\description{
Pretty much as it says, let's stage a base R \link{factor}.
}
\examples{
tmp <- tempfile()
dir.create(tmp)
stageObject(factor(1:10, 1:30), tmp, path="foo")
list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
