% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageBaseList.R
\name{stageObject,list-method}
\alias{stageObject,list-method}
\alias{stageObject,List-method}
\title{Stage a base list}
\usage{
\S4method{stageObject}{list}(x, dir, path, child = FALSE, fname = "list")

\S4method{stageObject}{List}(x, dir, path, child = FALSE, fname = "list")
}
\arguments{
\item{x}{A Bioconductor object of the specified class.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{fname}{String containing the name of the file to use to save \code{x}. 
Note that this should not have a \code{.json} suffix, so as to avoid confusion with the JSON-formatted metadata.}
}
\value{
A named list containing the metadata for \code{x}, where \code{x} itself is written to a JSON file.
}
\description{
Save a \link{list} or \linkS4class{List} to a JSON or HDF5 file, with external subdirectories created for any of the more complex list elements (e.g., DataFrames, arrays). 
This uses the uzuki2 specification to ensure that appropriate types are declared.
}
\section{File formats}{

If \code{\link{.saveBaseListFormat}() == "json.gz"}, the list is saved to a Gzip-compressed JSON file (the default).
This is an easily parsed format with low storage overhead.

If \code{\link{.saveBaseListFormat}() == "hdf5"}, \code{x} is saved into a HDF5 file instead.
This format is most useful for random access and for preserving the precision of numerical data.
}

\examples{
library(S4Vectors)
ll <- list(A=1, B=LETTERS, C=DataFrame(X=1:5))

tmp <- tempfile()
dir.create(tmp)
stageObject(ll, tmp, path="stuff")

list.files(tmp, recursive=TRUE)

}
\seealso{
\url{https://github.com/LTLA/uzuki2} for the specification.
}
\author{
Aaron Lun
}
