% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageDataFrameFactor.R
\name{stageObject,DataFrameFactor-method}
\alias{stageObject,DataFrameFactor-method}
\title{Stage a DataFrameFactor object}
\usage{
\S4method{stageObject}{DataFrameFactor}(
  x,
  dir,
  path,
  child = FALSE,
  index.name = "index",
  level.name = "levels",
  mcols.name = "mcols"
)
}
\arguments{
\item{x}{A \linkS4class{DataFrameFactor} object.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{index.name}{String containing the name of the file to save the factor indices.}

\item{level.name}{String containing the name of the subdirectory to save the factor levels.}

\item{mcols.name}{String specifying the name of the directory inside \code{path} to save the \code{\link{mcols}}.
If \code{NULL}, the metadata columns are not saved.}
}
\value{
A named list containing the metadata for \code{x}.
\code{x} itself is written to a file inside \code{path}.
Additional files may also be created inside \code{path} and referenced from the metadata.
}
\description{
Stage a \linkS4class{DataFrameFactor} object, a generalization of the base factor for \linkS4class{DataFrame} levels.
}
\details{
We create one file in \code{path} for the levels and another file for the factor indices. 
The levels file contains a DataFrame as staged by \code{\link{stageObject,DataFrame-method}}.
Indices are 1-based and reference one record of the levels file.
As the DataFrameFactor is a Vector subclass, its R-level metadata can be staged by \code{\link{.processMetadata}}.
}
\examples{
library(S4Vectors)
df <- DataFrame(X=LETTERS[1:5], Y=1:5)
out <- DataFrameFactor(df[sample(5, 100, replace=TRUE),,drop=FALSE])

tmp <- tempfile()
dir.create(tmp)
stageObject(out, tmp, path="test")

list.files(file.path(tmp, "test"), recursive=TRUE)

}
\seealso{
The \code{data_frame_factor} schema from \pkg{alabaster.schemas}.
}
\author{
Aaron Lun
}
