% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueValues.R
\name{.uniqueValues}
\alias{.uniqueValues}
\title{Getting unique values}
\usage{
.uniqueValues(x)
}
\arguments{
\item{x}{A character vector or \code{NULL}.}
}
\value{
The first 100 unique values from \code{x}, or \code{NULL} if \code{x=NULL}.
}
\description{
Get unique values, usually for the names or row names.
This is helpful for storing some values in the metadata for use in searching without creating excessively large documents.
}
\author{
Aaron Lun
}
