% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateObject.R
\name{validateObject}
\alias{validateObject}
\alias{validateObjectFunctionRegistry}
\alias{registerValidateObjectFunction}
\title{Validate an object's on-disk representation}
\usage{
validateObject(path, type = NULL, ...)

validateObjectFunctionRegistry()

registerValidateObjectFunction(type, fun)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with a \code{\link{saveObject}} method.}

\item{type}{String specifying the name of type of the object.
If this is not supplied for \code{readObject}, it is automatically determined from the \code{OBJECT} file in \code{path}.}

\item{fun}{A validation function that accepts \code{path} and \code{...}, and returns the associated object.
This may also be \code{NULL} to delete an existing registry.}
}
\value{
For \code{validateObject}, \code{NULL} is returned invisibly upon success, otherwise an error is raised.

For \code{validateObjectFunctionRegistry}, a named list of functions used to load each object type.

For \code{registerValidateObjectFunction}, the function is added to the registry.
}
\description{
Validate an object's on-disk representation.
This is done by dispatching to an appropriate validation function based on the type in the \code{OBJECT} file.
}
\examples{
library(S4Vectors)
df <- DataFrame(A=1:10, B=LETTERS[1:10])

tmp <- tempfile()
saveObject(df, tmp)
validateObject(tmp)

}
\author{
Aaron Lun
}
