% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vls.R
\name{vls}
\alias{vls}
\alias{h5_use_vls}
\alias{h5_write_vls_array}
\alias{h5_parse_vls_array}
\title{VLS saving utilities}
\description{
Utilities for saving our custom variable length string array format in HDF5.
Intended for \pkg{alabaster.*} developers only.
}
