% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{hdf5}
\alias{hdf5}
\alias{h5_guess_vector_chunks}
\alias{h5_create_vector}
\alias{h5_write_vector}
\alias{h5_write_attribute}
\alias{h5_read_vector}
\alias{h5_read_attribute}
\alias{h5_object_exists}
\alias{h5_cast}
\title{HDF5 utilities}
\description{
Basically just better versions of those in \pkg{rhdf5}, dedicated to \pkg{alabaster.base} and its dependents.
Intended for \pkg{alabaster.*} developers only.
}
