% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAtomicVector.R
\name{readAtomicVector}
\alias{readAtomicVector}
\alias{loadAtomicVector}
\title{Load an atomic vector}
\usage{
readAtomicVector(dir, path, ...)
}
\arguments{
\item{...}{Further arguments, ignored.}
}
\value{
The vector described by \code{info}, possibly with names.
}
\description{
Load a simple vector consisting of atomic elements from file.
}
\examples{
tmp <- tempfile()
dir.create(tmp)
meta <- stageObject(setNames(runif(26), letters), tmp, path="bar")
loadAtomicVector(meta, tmp)

}
\seealso{
\code{"\link{stageObject,integer-method}"}, for one of the staging methods.
}
\author{
Aaron Lun
}
