% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMetadata.R
\name{.processMcols}
\alias{.processMcols}
\alias{.processMetadata}
\title{Process R-level metadata}
\usage{
.processMcols(x, dir, path, mcols.name)

.processMetadata(x, dir, path, meta.name)
}
\arguments{
\item{x}{An \linkS4class{Vector} object for \code{.processMcols}, and a \linkS4class{Annotated} object for \code{.processMetadata}.}

\item{dir}{String containing the path to a staging directory.}

\item{path}{String containing the path within \code{dir} that is used to save all of \code{x}.
This should be the same as that passed to the \code{\link{stageObject}} method.}

\item{mcols.name}{String containing the name of the file to save \code{\link{mcols}(x)}.
If \code{NULL}, no saving is performed.}

\item{meta.name}{String containing the name of the file to save \code{\link{metadata}(x)}.
If \code{NULL}, no saving is performed.}
}
\value{
Both functions return a list containing \code{resource}, itself a list specifying the path within \code{dir} where the metadata was saved.
Alternatively \code{NULL} if no saving is performed.
}
\description{
Stage \code{\link{metadata}} and \code{\link{mcols}} for a \linkS4class{Annotated} or \linkS4class{Vector} object, respectively.
This is typically used inside \code{\link{stageObject}} methods for concrete subclasses.
}
\details{
If \code{mcols(x)} has no columns, nothing is saved by \code{.processMcols}.
Similarly, if \code{metadata(x)} is an empty list, nothing is saved by \code{.processMetadata}.
}
\seealso{
\code{.restoreMetadata}, which does the loading.
}
\author{
Aaron Lun
}
