% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAtomicVector.R
\name{readAtomicVector}
\alias{readAtomicVector}
\alias{loadAtomicVector}
\title{Read an atomic vector from disk}
\usage{
readAtomicVector(path, ...)
}
\arguments{
\item{path}{Path to a directory created with any of the vector methods for \code{\link{saveObject}}.}

\item{...}{Further arguments, ignored.}
}
\value{
The vector described by \code{info}, possibly with names.
}
\description{
Read a vector consisting of atomic elements from its on-disk representation.
}
\examples{
tmp <- tempfile()
saveObject(setNames(runif(26), letters), tmp)
readAtomicVector(tmp)

}
\seealso{
\code{"\link{saveObject,integer-method}"}, for one of the staging methods.
}
\author{
Aaron Lun
}
