% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBaseFactor.R
\name{readBaseFactor}
\alias{readBaseFactor}
\alias{loadBaseFactor}
\title{Read a factor from disk}
\usage{
readBaseFactor(path, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with the \code{\link{saveObject}} method for factors.}

\item{...}{Further arguments, ignored.}
}
\value{
The vector described by \code{info}, possibly with names.
}
\description{
Read a base R \link{factor} from its on-disk representation.
}
\examples{
tmp <- tempfile()
saveObject(factor(letters[1:10], letters), tmp)
readBaseFactor(tmp)

}
\seealso{
\code{"\link{stageObject,factor-method}"}, for the staging method.
}
\author{
Aaron Lun
}
