% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBaseFactor.R
\name{readBaseFactor}
\alias{readBaseFactor}
\alias{loadBaseFactor}
\alias{validateBaseFactor}
\title{Read a factor from disk}
\usage{
readBaseFactor(path, ...)

validateBaseFactor(path)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with the \code{\link{saveObject}} method for factors.}

\item{...}{Further arguments, ignored.}
}
\value{
For \code{readBaseFactor}, the vector described by \code{info}. 

For \code{validateBaseFactor}, an error is raised if the files are invalid.
Otherwise, \code{NULL} is invisibly returned.
}
\description{
Read a base R \link{factor} from its on-disk representation.
}
\examples{
tmp <- tempfile()
saveObject(factor(letters[1:10], letters), tmp)
readBaseFactor(tmp)

}
\seealso{
\code{"\link{saveObject,factor-method}"}, for the staging method.
}
\author{
Aaron Lun
}
