% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMetadata.R
\name{readMetadata}
\alias{readMetadata}
\alias{restoreMetadata}
\alias{.restoreMetadata}
\title{Read R-level metadata}
\usage{
readMetadata(x, metadata.path, mcols.path, ...)
}
\arguments{
\item{x}{An \linkS4class{Vector} or \linkS4class{Annotated} object.}

\item{metadata.path}{String containing a path to a directory, itself containing an on-disk representation of a base R list to be used as the \code{\link{metadata}}.}

\item{mcols.path}{String containing a path to a directory, itself containing an on-disk representation of a \linkS4class{DataFrame} to be used as the \code{\link{mcols}}.}

\item{...}{Further arguments to be passed to \code{\link{altReadObject}}.}
}
\value{
\code{x} is returned, possibly with \code{mcols} and \code{metadata} added to it.
}
\description{
Read \code{\link{metadata}} and \code{\link{mcols}} for a \linkS4class{Annotated} or \linkS4class{Vector} object, respectively.
This is typically used inside loading functions for concrete subclasses.
}
\seealso{
\code{\link{saveMetadata}}, which does the staging.
}
\author{
Aaron Lun
}
