% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageAtomicVector.R
\name{saveAtomicVector}
\alias{saveAtomicVector}
\alias{stageObject,integer-method}
\alias{stageObject,numeric-method}
\alias{stageObject,logical-method}
\alias{stageObject,character-method}
\alias{stageObject,double-method}
\alias{stageObject,POSIXct-method}
\alias{stageObject,Date-method}
\title{Stage atomic vectors}
\arguments{
\item{x}{Any of the atomic vector types, or \link{Date} objects, or time objects, e.g., \link{POSIXct}.}

\item{...}{Further arguments that are ignored.}
}
\value{
A named list containing the metadata for \code{x}.
\code{x} itself is written to a CSV file inside \code{path}.
}
\description{
Stage vectors containing atomic elements (or values that can be cast as such, e.g., dates and times).
}
\details{
Dates and POSIX times are cast to strings;
the type itself is recorded in the metadata.
}
\examples{
tmp <- tempfile()
dir.create(tmp)
stageObject(LETTERS, tmp, path="foo")
stageObject(setNames(runif(26), letters), tmp, path="bar")

list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
